+++ jmx/jmx-command-client/src/test/java/uk/gov/justice/services/jmx/system/command/client/SystemCommanderClientFactoryTest.java
@@ -1,18 +1,15 @@
 package uk.gov.justice.services.jmx.system.command.client;
 
-import static com.google.common.collect.ImmutableMap.of;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
+import static uk.gov.justice.services.jmx.system.command.client.connection.JmxParametersBuilder.jmxParameters;
 
 import uk.gov.justice.services.jmx.system.command.client.build.ObjectFactory;
-import uk.gov.justice.services.jmx.system.command.client.connection.CredentialsFactory;
 import uk.gov.justice.services.jmx.system.command.client.connection.JMXConnectorFactory;
+import uk.gov.justice.services.jmx.system.command.client.connection.JmxParametersBuilder;
 import uk.gov.justice.services.jmx.system.command.client.connection.MBeanConnector;
-
-import java.util.HashMap;
-import java.util.Map;
 
 import javax.management.remote.JMXConnector;
 
@@ -37,6 +34,8 @@ public class SystemCommanderClientFactoryTest {
         final String hostName = "localhost";
         final int port = 9340;
 
+        final JmxParametersBuilder jmxParametersBuilder = jmxParameters().withHost(hostName).withPort(port);
+
         final MBeanConnector mBeanConnector = mock(MBeanConnector.class);
         final JMXConnectorFactory jmxConnectorFactory = mock(JMXConnectorFactory.class);
         final JMXConnector jmxConnector = mock(JMXConnector.class);
@@ -44,34 +43,9 @@ public class SystemCommanderClientFactoryTest {
 
         when(objectFactory.mBeanConnector()).thenReturn(mBeanConnector);
         when(objectFactory.jmxConnectorFactory()).thenReturn(jmxConnectorFactory);
-        when(jmxConnectorFactory.createJmxConnector(hostName, port, new HashMap<>())).thenReturn(jmxConnector);
+        when(jmxConnectorFactory.createJmxConnector(jmxParametersBuilder)).thenReturn(jmxConnector);
         when(objectFactory.systemCommanderClient(mBeanConnector, jmxConnector)).thenReturn(systemCommanderClient);
 
-        assertThat(systemCommanderClientFactory.create(hostName, port), is(systemCommanderClient));
-    }
-
-    @Test
-    public void shouldCreateSystemCommanderClientWithCredentials() throws Exception {
-
-        final String hostName = "localhost";
-        final int port = 9340;
-        final String username = "Fred";
-        final String password = "Password123";
-        final Map<String, Object> environmentWithCredentials = of(username, password);
-
-        final MBeanConnector mBeanConnector = mock(MBeanConnector.class);
-        final JMXConnectorFactory jmxConnectorFactory = mock(JMXConnectorFactory.class);
-        final JMXConnector jmxConnector = mock(JMXConnector.class);
-        final SystemCommanderClient systemCommanderClient = mock(SystemCommanderClient.class);
-        final CredentialsFactory credentialsFactory = mock(CredentialsFactory.class);
-
-        when(objectFactory.mBeanConnector()).thenReturn(mBeanConnector);
-        when(objectFactory.jmxConnectorFactory()).thenReturn(jmxConnectorFactory);
-        when(objectFactory.credentialsFactory()).thenReturn(credentialsFactory);
-        when(credentialsFactory.create(username, password)).thenReturn(environmentWithCredentials);
-        when(jmxConnectorFactory.createJmxConnector(hostName, port, environmentWithCredentials)).thenReturn(jmxConnector);
-        when(objectFactory.systemCommanderClient(mBeanConnector, jmxConnector)).thenReturn(systemCommanderClient);
-
-        assertThat(systemCommanderClientFactory.create(hostName, port, username, password), is(systemCommanderClient));
+        assertThat(systemCommanderClientFactory.create(jmxParametersBuilder), is(systemCommanderClient));
     }
 }